+++ changelog.md
@@ -36,6 +36,7 @@ Codion Change Log
 - EntityTablePanel, export control moved out of popup copy submenu, icon added.
 - FrameworkMessages.copyTableWithHeader() removed, moved to EntityTablePanel i18n messages.
 - EntityApplicationPanel, modified warning improved, message now includes modified attributes, panel activated and modified input field focused.
+- EntityEditPanel, modified warning improved, message now includes modified attributes, modified input field focused.
 
 ## 0.18.62
 ### is.codion.common.reactive